added comprobations + fixes + volume
for when the user calling some commands that affect the bot (play, leave, skip, etc) is not in the same voice channel/not in a voice channel.
""fixed"" bot cutting. Apparently from restarting the bot so much it created multiple instances or something and terminated the session, but thats just for me. Make sure to not restart the bot very often and if possible wait for it to disconnect completely
added basic volume control with 'volume' command, given a % of volume.
videos now properly get and keep their volume/pitch/speed data when changing with goto/remove/skip/rewind, and changed pitch/speed method to ffmpeg's "rubberband"
fixed some minor bugs